+++ script/ci/download.sh
@@ -1,0 +1,1 @@
+node script/ci/download.js